expanded readme and made module suitable for python use

+++ cythonbuilder/cython_builder.py
@@ -12,19 +12,18 @@ project_dir = os.getcwd()
 
 
 
-def init_cython():
+def cy_init():
     """ Creates projdir.ext, projdir.ext.annotations """
 
     # create root.ext, root.ext.annotations
     FilesAndFolders.create_folder(folderpath=os.path.join(project_dir, appsettings.cython_extensions_dirname))
     FilesAndFolders.create_folder(folderpath=os.path.join(project_dir, appsettings.cython_anno_dirname))
+    logger.debug(msg=f"Initialized cybuider at {project_dir}")
 
 
 
-
-def cython_find_pyx_files(target_files:[str]=None) -> [str]:
+def cy_list(target_files:[str]=None) -> [str]:
     """ Target files is optional filter. Returns a list of fullpaths to pyxfiles """
-
     # 1. Find all fullpaths to pyx files in all folders but the /venv
     venv_paths_folders = [os.path.dirname(path) for path in Path(project_dir).rglob('pyvenv.cfg')]
     pyx_fullpaths = [str(pyxpath) for pyxpath in Path(project_dir).rglob('*.pyx')]
@@ -40,8 +39,8 @@ def cython_find_pyx_files(target_files:[str]=None) -> [str]:
         diff = set(target_file_names).difference(set(my_pyx_file_names))
         intersection = set(my_pyx_file_names).intersection(set(target_file_names))
 
-        if (len(diff) > 0):
-            raise ValueError(f"Cannot find provided target files: {[f'{f}.pyx' for f in diff]}")
+        # if (len(diff) > 0):
+        #     raise ValueError(f"Cannot find provided target files: {[f'{f}.pyx' for f in diff]}")
 
         _pyx_fullpaths = []
         for pyx in pyx_fullpaths:
@@ -50,9 +49,11 @@ def cython_find_pyx_files(target_files:[str]=None) -> [str]:
                     _pyx_fullpaths.append(pyx)
         pyx_fullpaths = _pyx_fullpaths
     return pyx_fullpaths
-def cython_build(target_files:[str], create_annotations:bool=True, include_numpy:bool=True):
+def cy_build(target_files:[str] = None, create_annotations:bool=True, include_numpy:bool=False):
     """ Builds all pyx files in the /ext folder """
 
+    if (target_files == None):
+        target_files = cy_list()
 
     # We want to build in place
     sys.argv = [sys.argv[0], 'build_ext', '--inplace']
@@ -68,6 +69,9 @@ def cython_build(target_files:[str], create_annotations:bool=True, include_numpy
     # Create Extension objects
     ext_modules = []
     for n in target_files:
+        if (not os.path.isfile(n)):
+            logger.info(msg=f"File {n} not found; skipping..")
+            continue
         logger.debug(msg=f"C {n}")
         module_name, extension = os.path.splitext(os.path.basename(n))
         # The name must be plain, no path
@@ -94,17 +98,25 @@ def cython_build(target_files:[str], create_annotations:bool=True, include_numpy
         ext_modules=cythonize(ext_modules),
         # buid_dir=path_build_dir
     )
-def cython_clean(target_files:[str], keep_c_files:bool=False):
+def cy_clean(target_files:[str] = None, keep_c_files:bool=False):
     """ Clean up all files """
-
     logger.debug(msg="cleanup")
+
+    cy_init()
+    if (target_files == None):
+        target_files = cy_list()
+
     annotations_dir = os.path.join(project_dir, appsettings.cython_anno_dirname)
 
     # Remove build folder
     build_folder_path = os.path.join(project_dir, 'build')
-    shutil.rmtree(path=build_folder_path)
+    if (os.path.isdir(build_folder_path)):
+        shutil.rmtree(path=build_folder_path)
 
     for built_file in target_files:
+        if (not os.path.isfile(built_file)):
+            logger.info(msg=f"File {built_file} not found; skipping..")
+            continue
 
         _filename =os.path.splitext(os.path.basename(built_file))[0]    # no ext
